Kafka output encode parameter
