+++ plugin/output/http/http_test.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/ozontech/file.d/encoder"
 	"github.com/ozontech/file.d/pipeline"
 	"github.com/ozontech/file.d/test"
 	insaneJSON "github.com/ozontech/insane-json"
@@ -26,16 +27,18 @@ func TestEncoding(t *testing.T) {
 	data := data{}
 	event := &pipeline.Event{Root: root}
 
-	encoder := newJSONEncoder(&JSONEncoderParams{})
-	data.outBuf = encoder.Encode(event, data.outBuf)
+	jsonEncoder, _ := encoder.NewEncoder(encoder.EncodingConfig{
+		Type: encoder.EncoderTypeJSON,
+	})
+	data.outBuf = jsonEncoder.Encode(event, data.outBuf)
 	data.outBuf = append(data.outBuf, '\n')
 
 	expected := fmt.Sprintf("%s\n", `{"message":"[INFO] some event","field_a":"AAAA","field_b":"BBBB"}`)
 	assert.Equal(t, expected, string(data.outBuf), "wrong request content")
 
-	var params RawEncoderParams
-	rawEncoder := newRawEncoder(&params)
-
+	rawEncoder, _ := encoder.NewEncoder(encoder.EncodingConfig{
+		Type: encoder.EncoderTypeRaw,
+	})
 	data.outBuf = data.outBuf[:0]
 	data.outBuf = rawEncoder.Encode(event, data.outBuf)
 	data.outBuf = append(data.outBuf, '\n')
